lab-3/pyproject.toml
[project]
name = "lab-3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pycryptodome>=3.23.0",
]
